chore: rename Vue components
